requirements.txt
# Backend
fastapi
uvicorn

# Gemini AI
google-generativeai

# Document engine
llama-index
llama-index-llms-gemini

# Utilities
python-dotenv
requests
